Created template for home page and contract brewery home page
