+++ frontend/brew_it/lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:brew_it/core/theme/theme_constants.dart';
-import 'package:brew_it/presentation/home/home_page.dart';
+import 'package:brew_it/presentation/home/home_page_commercial.dart';
+import 'package:brew_it/presentation/home/home_page_contract.dart';
 import 'package:flutter/material.dart';
 
 void main() {
@@ -14,7 +15,7 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       title: 'BrewIT Demo',
       theme: baseTheme,
-      home: const HomePage(),
+      home: HomePageContract(),
     );
   }
 }
