+++ frontend/brew_it/lib/presentation/home/home_page_template.dart
@@ -1,0 +1,39 @@
+import 'package:brew_it/presentation/common/menu_button.dart';
+import 'package:flutter/material.dart';
+
+class HomePageTemplate extends StatelessWidget {
+  const HomePageTemplate(
+      {required this.title, required this.buttons, super.key});
+
+  final String title;
+  final List<MenuButton> buttons;
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+        appBar: AppBar(),
+        body: Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
+          Expanded(flex: 2, child: Container()),
+          Expanded(
+            flex: 6,
+            child: Padding(
+              padding: const EdgeInsets.only(left: 60),
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: <Widget>[
+                      Text(
+                        title,
+                        style: const TextStyle(
+                          fontWeight: FontWeight.bold,
+                        ),
+                      ),
+                    ] +
+                    buttons,
+              ),
+            ),
+          ),
+          Expanded(flex: 2, child: Container())
+        ]));
+  }
+}
